Experimental Design Capstone Presentation
These are the slides created in order to present the results from the data analysis for my Experimental Design Capstone.

Please Note: If you are looking through these slides without context, there are animations that perhaps make it seem as if the slides are disorganized and don't make sense. This presentation is intended to be experienced live.